Moved Scenarios to its own service, moved things to other services, cleanups and removed deprecated things
updates

moved scenarios to its own service

moved more things to individual services

moved types to scenario types, removed dependency from treatment goals in scenario service

refactor code into PlanStateService

removed unused stuff about configs

remove deprecated updateStateWithConfig
