+++ src/interface/src/app/map/map.component.spec.ts
@@ -25,7 +25,6 @@ import {
   AuthService,
   MapService,
   PlanService,
-  PlanState,
   SessionService,
 } from '../services';
 import {
@@ -47,6 +46,7 @@ import { PlanCreateDialogComponent } from './plan-create-dialog/plan-create-dial
 import { ProjectCardComponent } from './project-card/project-card.component';
 import { SignInDialogComponent } from './sign-in-dialog/sign-in-dialog.component';
 import { FeaturesModule } from '../features/features.module';
+import { PlanState, PlanStateService } from '../services/plan-state.service';
 
 describe('MapComponent', () => {
   let component: MapComponent;
@@ -133,7 +133,7 @@ describe('MapComponent', () => {
         }),
       }
     );
-    const fakePlanService = jasmine.createSpyObj<PlanService>(
+    const fakePlanStateService = jasmine.createSpyObj<PlanStateService>(
       'PlanService',
       { createPlan: of({ success: true, result: fakePlan }) },
       {
@@ -141,7 +141,6 @@ describe('MapComponent', () => {
           all: {}, // All plans indexed by id
           currentPlanId: 'temp',
           currentScenarioId: null,
-          currentConfigId: null,
           mapConditionLayer: null,
           mapShapes: null,
           legendUnits: null,
@@ -188,7 +187,7 @@ describe('MapComponent', () => {
         { provide: AuthService, useValue: fakeAuthService },
         { provide: MatDialog, useValue: fakeMatDialog },
         { provide: MapService, useValue: fakeMapService },
-        { provide: PlanService, useValue: fakePlanService },
+        { provide: PlanService, useValue: fakePlanStateService },
         { provide: SessionService, useValue: fakeSessionService },
         { provide: Router, useFactory: routerStub },
       ],
