+++ src/interface/src/app/plan/create-scenarios/create-scenarios.component.spec.ts
@@ -9,7 +9,7 @@ import {
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BehaviorSubject, of } from 'rxjs';
-import { PlanService, PlanState } from 'src/app/services';
+
 import {
   Region,
   Scenario,
@@ -23,6 +23,8 @@ import { HarnessLoader } from '@angular/cdk/testing';
 import { MatButtonHarness } from '@angular/material/button/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { POLLING_INTERVAL } from '../plan-helpers';
+import { ScenarioService } from '../../services/scenario.service';
+import { PlanState, PlanStateService } from '../../services/plan-state.service';
 
 //TODO Add the following tests once implementation for tested behaviors is added:
 /**
@@ -34,7 +36,8 @@ import { POLLING_INTERVAL } from '../plan-helpers';
 describe('CreateScenariosComponent', () => {
   let component: CreateScenariosComponent;
   let fixture: ComponentFixture<CreateScenariosComponent>;
-  let fakePlanService: PlanService;
+  let fakePlanService: PlanStateService;
+  let fakeScenarioService: ScenarioService;
 
   let loader: HarnessLoader;
   let defaultSelectedQuestion: TreatmentQuestionConfig = {
@@ -52,20 +55,16 @@ describe('CreateScenariosComponent', () => {
   };
 
   beforeEach(async () => {
-    fakePlanService = jasmine.createSpyObj<PlanService>(
-      'PlanService',
+    fakeScenarioService = jasmine.createSpyObj<ScenarioService>(
+      'ScenarioService',
       {
-        getConditionScoresForPlanningArea: of(),
-        getProject: of({
-          id: 1,
-          maxBudget: 100,
-        }),
-        updateProject: of(1),
-        createProjectArea: of(1),
         createScenario: of('1'),
         getScenario: of(fakeScenario),
-        updateStateWithShapes: undefined,
-      },
+      }
+    );
+    fakePlanService = jasmine.createSpyObj<PlanStateService>(
+      'PlanStateService',
+      {},
       {
         planState$: new BehaviorSubject<PlanState>({
           all: {
@@ -77,7 +76,6 @@ describe('CreateScenariosComponent', () => {
             },
           },
           currentPlanId: '1',
-          currentConfigId: 1,
           currentScenarioId: '1',
           mapConditionLayer: null,
           mapShapes: null,
@@ -112,7 +110,10 @@ describe('CreateScenariosComponent', () => {
         RouterTestingModule,
       ],
       declarations: [CreateScenariosComponent],
-      providers: [{ provide: PlanService, useValue: fakePlanService }],
+      providers: [
+        { provide: PlanStateService, useValue: fakePlanService },
+        { provide: ScenarioService, useValue: fakeScenarioService },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(CreateScenariosComponent);
@@ -129,7 +130,7 @@ describe('CreateScenariosComponent', () => {
   it('should load existing scenario', () => {
     spyOn(component, 'pollForChanges');
     fixture.detectChanges();
-    expect(fakePlanService.getScenario).toHaveBeenCalledOnceWith('1');
+    expect(fakeScenarioService.getScenario).toHaveBeenCalledOnceWith('1');
     component.formGroups[2].valueChanges.subscribe((_) => {
       expect(component.formGroups[2].get('budgetForm.maxCost')?.value).toEqual(
         100
