+++ src/interface/src/app/plan/create-scenarios/set-priorities/set-priorities.component.spec.ts
@@ -7,7 +7,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BehaviorSubject, of } from 'rxjs';
 import { MaterialModule } from 'src/app/material/material.module';
 import { SharedModule } from 'src/app/shared/shared.module';
-import { TreatmentGoalConfig, TreatmentQuestionConfig } from 'src/app/types';
+import { TreatmentQuestionConfig } from 'src/app/types';
 
 import { MapService } from './../../../services/map.service';
 import { PlanService } from './../../../services/plan.service';
@@ -85,37 +85,24 @@ describe('SetPrioritiesComponent', () => {
         }),
       }
     );
-    fakePlanService = jasmine.createSpyObj<PlanService>(
-      'PlanService',
-      {
-        getConditionScoresForPlanningArea: of({
-          conditions: [
-            {
-              condition: 'test_pillar_1',
-              mean_score: 0.1,
-            },
-            {
-              condition: 'test_element_1',
-              mean_score: -0.7,
-            },
-            {
-              condition: 'test_metric_1',
-              mean_score: 0.4,
-            },
-          ],
-        }),
-      },
-      {
-        treatmentGoalsConfig$: new BehaviorSubject<
-          TreatmentGoalConfig[] | null
-        >([
+    fakePlanService = jasmine.createSpyObj<PlanService>('PlanService', {
+      getConditionScoresForPlanningArea: of({
+        conditions: [
           {
-            category_name: 'test_category',
-            questions: [testQuestion],
+            condition: 'test_pillar_1',
+            mean_score: 0.1,
           },
-        ]),
-      }
-    );
+          {
+            condition: 'test_element_1',
+            mean_score: -0.7,
+          },
+          {
+            condition: 'test_metric_1',
+            mean_score: 0.4,
+          },
+        ],
+      }),
+    });
     await TestBed.configureTestingModule({
       imports: [
         BrowserAnimationsModule,
