--- src/interface/src/app/plan/plan-summary/scenario-configurations/scenario-configurations.component.html
@@ -1,118 +1,0 @@
-<div class="configurations-wrapper">
-  <mat-card [style.margin-top]="'24px'">
-    <mat-card-header>
-      <mat-card-title>
-        My Configurations ({{ configurations.length }})
-      </mat-card-title>
-    </mat-card-header>
-
-    <mat-card-content>
-      <p>{{ text1 }}</p>
-
-      <mat-table [dataSource]="configurations">
-        <!-- Select Column -->
-        <ng-container matColumnDef="select">
-          <mat-header-cell *matHeaderCellDef>
-            <mat-checkbox (change)="selectAllConfigs($event)"></mat-checkbox>
-          </mat-header-cell>
-          <mat-cell *matCellDef="let element">
-            <div class="hover-indicator"></div>
-            <mat-checkbox
-              [(ngModel)]="element.selected"
-              (click)="$event.stopPropagation()"></mat-checkbox>
-          </mat-cell>
-        </ng-container>
-
-        <!-- Timestamp Column -->
-        <ng-container matColumnDef="createdTimestamp">
-          <mat-header-cell *matHeaderCellDef>Date Created</mat-header-cell>
-          <mat-cell *matCellDef="let element">
-            {{ element.createdTimestamp | date: 'medium' }}
-          </mat-cell>
-        </ng-container>
-
-        <!-- Score Type Column -->
-        <ng-container matColumnDef="scoreType">
-          <mat-header-cell *matHeaderCellDef>Calculating Score</mat-header-cell>
-          <mat-cell *matCellDef="let element">Condition Score</mat-cell>
-        </ng-container>
-
-        <!-- Priorities Column -->
-        <ng-container matColumnDef="priorities">
-          <mat-header-cell *matHeaderCellDef>
-            Selected Priorities
-          </mat-header-cell>
-          <mat-cell *matCellDef="let element">
-            <div>
-              <div
-                *ngFor="let priority of displayPriorities(element.priorities)">
-                {{ priority }}
-              </div>
-            </div>
-          </mat-cell>
-        </ng-container>
-
-        <!-- Constraints Column -->
-        <ng-container matColumnDef="constraints">
-          <mat-header-cell *matHeaderCellDef>
-            Selected Constraints
-          </mat-header-cell>
-          <mat-cell *matCellDef="let element">
-            <div>
-              <div *ngIf="element.max_budget">
-                Max budget: ${{ element.max_budget }}
-              </div>
-              <div *ngIf="element.max_treatment_area_ratio">
-                Max treatment area: {{ element.max_treatment_area_ratio }}%
-              </div>
-              <div *ngIf="element.max_slope">
-                Exclude slope > {{ element.max_slope }} degrees
-              </div>
-              <div *ngIf="element.min_distance_from_road">
-                Exclude areas off road by
-                {{ element.min_distance_from_road }} ft.
-              </div>
-            </div>
-          </mat-cell>
-        </ng-container>
-
-        <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
-        <mat-row
-          *matRowDef="let row; columns: displayedColumns"
-          class="config-row"
-          (click)="openConfig(row)"></mat-row>
-      </mat-table>
-
-      <!-- Padding for visual effect if the table is empty -->
-      <div class="no-configs-padding" *ngIf="configurations.length === 0"></div>
-    </mat-card-content>
-
-    <mat-card-actions align="end">
-      <button
-        mat-button
-        [disabled]="!showDeleteButton()"
-        (click)="deleteSelectedConfigs()">
-        <mat-icon>delete</mat-icon>
-        DELETE
-      </button>
-      <button
-        mat-raised-button
-        [disabled]="!showContinueButton()"
-        color="primary"
-        (click)="openConfig()">
-        CONTINUE PLANNING
-      </button>
-    </mat-card-actions>
-  </mat-card>
-
-  <div *ngIf="configurations.length === 0" class="no-configs-overlay">
-    <p>
-      Configurations use priorities & constraints to identify and rank project
-      areas for treatment.
-    </p>
-    <button mat-raised-button color="primary" (click)="openConfig()">
-      <mat-icon class="material-symbols-outlined">add_box</mat-icon>
-      NEW CONFIGURATION
-    </button>
-  </div>
-</div>
