--- src/interface/src/app/plan/plan-summary/scenario-configurations/scenario-configurations.component.scss
@@ -1,101 +1,0 @@
-@use 'sass:map';
-@use '@angular/material' as mat;
-
-@import '../../../../theme.scss';
-
-$color-config: mat.get-color-config($planscape-frontend-theme);
-$primary-palette: map.get($color-config, 'primary');
-
-.configurations-wrapper {
-  position: relative;
-}
-
-.mat-card {
-  border-radius: 16px;
-}
-
-.mat-card-content {
-  padding: 8px;
-}
-
-.mat-card-actions {
-  gap: 12px;
-}
-
-.no-configs-padding {
-  height: 50px;
-}
-
-.no-configs-overlay {
-  align-items: center;
-  backdrop-filter: blur(2px);
-  background: rgba(79, 79, 79, 0.8);
-  border-radius: 16px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: center;
-  padding: 24px;
-  position: absolute;
-  top: 0;
-  width: 100%;
-  z-index: 2;
-
-  p {
-    color: white;
-    font-size: 18px;
-    font-weight: 400;
-    line-height: 24px;
-    text-align: center;
-  }
-}
-
-.mat-header-cell {
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 18px;
-}
-
-.mat-cell {
-  font-size: 13px;
-  font-weight: 400;
-  line-height: 18px;
-}
-
-.mat-column-select {
-  flex: 0 1 38px;
-  padding-left: 16px;
-  padding-right: 32px;
-  width: 38px;
-
-  &.mat-header-cell {
-    padding-left: 22px;
-  }
-}
-
-.mat-column-constraints.mat-cell {
-  opacity: 0.66;
-}
-
-.config-row {
-  cursor: pointer;
-
-  &:hover {
-    background-color: transparentize(
-      $color: mat.get-color-from-palette($primary-palette, 500),
-      $amount: 0.9
-    );
-
-    mat-cell div.hover-indicator {
-      background-color: mat.get-color-from-palette($primary-palette, 500);
-    }
-  }
-}
-
-.hover-indicator {
-  height: 48px;
-  margin-left: -16px;
-  margin-right: 18px;
-  width: 4px;
-}
