+++ src/interface/src/app/services/errors.ts
@@ -1,0 +1,5 @@
+export class CreateScenarioError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
