+++ src/interface/src/app/services/plan.service.spec.ts
@@ -5,16 +5,17 @@ import {
 import { TestBed } from '@angular/core/testing';
 
 import { BackendConstants } from '../backend-constants';
-import { BasePlan, Plan, Region } from '../types';
 import {
-  PlanPreview,
-  Scenario,
-  ScenarioConfig,
-  TreatmentGoalConfig,
-  TreatmentQuestionConfig,
-} from './../types/plan.types';
-import { BackendPlan, BackendPlanPreview, PlanService } from './plan.service';
+  BackendPlan,
+  BackendPlanPreview,
+  BasePlan,
+  Plan,
+  Region,
+} from '../types';
+import { PlanPreview } from './../types/plan.types';
+import { PlanService } from './plan.service';
 import { MapService } from './map.service';
+
 // TODO Make test for getting scenario results
 // TODO Make test for call to create project area
 // TODO Make test for call to bulk create project area
@@ -26,7 +27,6 @@ describe('PlanService', () => {
   let mockPlan: BasePlan;
   let fakeGeoJson: GeoJSON.GeoJSON;
 
-  const treatmentGoalConfigs: TreatmentGoalConfig[] = [];
   beforeEach(() => {
     fakeGeoJson = {
       type: 'FeatureCollection',
@@ -44,13 +44,6 @@ describe('PlanService', () => {
     });
     service = TestBed.inject(PlanService);
     httpTestingController = TestBed.inject(HttpTestingController);
-    // Must flush the requests in the constructor for httpTestingController.verify()
-    // to pass in other tests.
-    const req1 = httpTestingController.expectOne(
-      BackendConstants.END_POINT +
-        '/planning/treatment_goals_config/?region_name=sierra-nevada'
-    );
-    req1.flush(treatmentGoalConfigs);
   });
 
   describe('deletePlan', () => {
@@ -163,103 +156,6 @@ describe('PlanService', () => {
   });
 
   // TODO Update once Project/Scenario types are updated
-  describe('getScenario', () => {
-    it('should make HTTP request to backend', () => {
-      const scenarioConfig: ScenarioConfig = {
-        est_cost: undefined,
-        max_budget: undefined,
-        min_distance_from_road: undefined,
-        max_slope: undefined,
-        max_treatment_area_ratio: undefined,
-        treatment_question: null,
-        created_timestamp: undefined,
-        project_areas: [],
-        excluded_areas: undefined,
-        stand_size: 'MEDIUM',
-      };
-      const scenario: Scenario = {
-        id: '1',
-        name: 'name',
-        notes: undefined,
-        planning_area: '1',
-        configuration: scenarioConfig,
-        scenario_result: undefined,
-      };
-
-      service.getScenario('1').subscribe((res) => {
-        expect(res).toEqual(scenario);
-      });
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/planning/get_scenario_by_id/?id=1')
-      );
-      expect(req.request.method).toEqual('GET');
-      req.flush(scenario);
-    });
-  });
-
-  describe('createScenario', () => {
-    let defaultSelectedQuestion: TreatmentQuestionConfig = {
-      short_question_text: '',
-      long_question_text: '',
-      scenario_output_fields_paths: { metrics: [''] },
-      scenario_priorities: [''],
-      stand_thresholds: [''],
-      global_thresholds: [''],
-      weights: [0],
-    };
-    it('should make HTTP request to backend', (done) => {
-      const scenarioConfig: ScenarioConfig = {
-        est_cost: 1,
-        min_distance_from_road: 1,
-        max_slope: 1,
-        max_treatment_area_ratio: 1,
-        excluded_areas: undefined,
-        max_budget: 200,
-        treatment_question: defaultSelectedQuestion,
-        stand_size: 'LARGE',
-      };
-      const scenario: Scenario = {
-        name: 'name',
-        planning_area: '1',
-        configuration: scenarioConfig,
-      };
-
-      const backendConfig = {
-        est_cost: 1,
-        min_distance_from_road: 1,
-        max_slope: 1,
-        max_treatment_area_ratio: 1,
-        excluded_areas: undefined,
-        max_budget: 200,
-        scenario_output_fields: [''],
-        scenario_priorities: [''],
-        stand_thresholds: [''],
-        global_thresholds: [''],
-        weights: [0],
-        stand_size: 'LARGE',
-      };
-
-      service.createScenario(scenario).subscribe((res) => {
-        expect(res).toEqual('1');
-        done();
-      });
-
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/planning/create_scenario/')
-      );
-      expect(req.request.method).toEqual('POST');
-
-      expect(req.request.body).toEqual({
-        name: 'name',
-        planning_area: '1',
-        configuration: backendConfig,
-      });
-      req.flush('1');
-      httpTestingController.verify();
-    });
-  });
-
-  // TODO Update once Project/Scenario types are updated
   // describe('getScenariosForPlan', () => {
   //   it('should make HTTP request to backend', (done) => {
   //     const projectConfig: ProjectConfig = {
@@ -308,90 +204,4 @@ describe('PlanService', () => {
   //     httpTestingController.verify();
   //   });
   // });
-
-  describe('updateScenarioNotes', () => {
-    it('should make HTTP request to backend', () => {
-      const scenarioConfig: ScenarioConfig = {
-        est_cost: undefined,
-        max_budget: undefined,
-        min_distance_from_road: undefined,
-        max_slope: undefined,
-        max_treatment_area_ratio: undefined,
-        created_timestamp: undefined,
-        treatment_question: undefined,
-      };
-      const scenario: Scenario = {
-        id: '1',
-        name: 'name',
-        planning_area: '1',
-        configuration: scenarioConfig,
-        notes: 'hello',
-      };
-
-      service.updateScenarioNotes(scenario).subscribe((res) => {
-        expect(res).toEqual(1);
-      });
-
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/planning/update_scenario/')
-      );
-      expect(req.request.body).toEqual({
-        id: scenario.id,
-        notes: scenario.notes,
-      });
-      expect(req.request.method).toEqual('PATCH');
-      req.flush(1);
-      httpTestingController.verify();
-    });
-  });
-
-  describe('deleteScenarios', () => {
-    it('should make HTTP request to backend', (done) => {
-      service.deleteScenarios(['1']).subscribe((res) => {
-        expect(res).toEqual(['1']);
-        done();
-      });
-
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/planning/delete_scenario/')
-      );
-      expect(req.request.method).toEqual('POST');
-      req.flush(['1']);
-      httpTestingController.verify();
-    });
-  });
-
-  describe('favoriteScenario', () => {
-    it('should make HTTP request to backend', (done) => {
-      service.favoriteScenario('1').subscribe((res) => {
-        expect(res).toEqual({ favorited: true });
-        done();
-      });
-
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/plan/favorite_scenario/')
-      );
-      expect(req.request.method).toEqual('POST');
-      expect(req.request.body).toEqual({ scenario_id: 1 });
-      req.flush({ favorited: true });
-      httpTestingController.verify();
-    });
-  });
-
-  describe('unfavoriteScenario', () => {
-    it('should make HTTP request to backend', (done) => {
-      service.unfavoriteScenario('1').subscribe((res) => {
-        expect(res).toEqual({ favorited: false });
-        done();
-      });
-
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/plan/unfavorite_scenario/')
-      );
-      expect(req.request.method).toEqual('POST');
-      expect(req.request.body).toEqual({ scenario_id: 1 });
-      req.flush({ favorited: false });
-      httpTestingController.verify();
-    });
-  });
 });
