+++ src/interface/src/app/services/plan.service.ts
@@ -1,107 +1,22 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import {
-  BehaviorSubject,
-  catchError,
-  map,
-  Observable,
-  of,
-  take,
-  tap,
-} from 'rxjs';
+import { map, Observable, take } from 'rxjs';
 
 import { BackendConstants } from '../backend-constants';
-import { BasePlan, Plan, Region, regionToString } from '../types';
 import {
-  Scenario,
-  ScenarioConfig,
+  BackendPlan,
+  BackendPlanPreview,
+  BasePlan,
+  Plan,
   PlanConditionScores,
-  PlanPreview,
-  ProjectArea,
-  ProjectConfig,
-  TreatmentGoalConfig,
-  TreatmentQuestionConfig,
-} from './../types/plan.types';
-
-// TODO Remove Config
-export interface PlanState {
-  all: {
-    [planId: string]: Plan;
-  };
-  currentPlanId: Plan['id'] | null;
-  currentScenarioId: Scenario['id'] | null;
-  currentConfigId: ProjectConfig['id'] | null;
-  mapConditionLayer: string | null;
-  mapShapes: any | null;
-  legendUnits: string | null;
-}
-
-export interface BackendPlan {
-  id?: number;
-  name: string;
-  user?: number;
-  notes?: string;
-  region_name: Region;
-  geometry?: GeoJSON.GeoJSON;
-  scenario_count?: number;
-  projects?: number;
-  created_at?: string;
-  latest_updated?: string;
-}
-
-export interface BackendPlanPreview {
-  id: number;
-  name: string;
-  notes: string;
-  user: number;
-  region_name: Region;
-  scenario_count: number;
-  latest_updated: string;
-  geometry?: GeoJSON.GeoJSON;
-}
-
-export interface BackendProjectArea {
-  id: number;
-  geometry: GeoJSON.GeoJSON;
-  properties?: {
-    estimated_area_treated?: number;
-    owner?: number;
-    project?: number;
-  };
-}
+} from '../types';
+import { PlanPreview } from './../types/plan.types';
 
 @Injectable({
   providedIn: 'root',
 })
 export class PlanService {
-  // Warning: do not mutate state!
-  readonly planState$ = new BehaviorSubject<PlanState>({
-    all: {}, // All plans indexed by id
-    currentPlanId: null,
-    currentScenarioId: null,
-    currentConfigId: null,
-    mapConditionLayer: null,
-    mapShapes: null,
-    legendUnits: null,
-  });
-  readonly treatmentGoalsConfig$ = new BehaviorSubject<
-    TreatmentGoalConfig[] | null
-  >(null);
-
-  readonly planRegion$ = new BehaviorSubject<Region>(Region.SIERRA_NEVADA);
-
-  constructor(private http: HttpClient) {
-    this.http
-      .get<TreatmentGoalConfig[]>(
-        BackendConstants.END_POINT +
-          '/planning/treatment_goals_config/?region_name=' +
-          `${regionToString(this.planRegion$.getValue())}`
-      )
-      .pipe(take(1))
-      .subscribe((config: TreatmentGoalConfig[]) => {
-        this.treatmentGoalsConfig$.next(config);
-      });
-  }
+  constructor(private http: HttpClient) {}
 
   planNameExists(planName: string) {
     return this.listPlansByUser().pipe(
@@ -115,9 +30,10 @@ export class PlanService {
    * */
 
   /** Makes a request to the backend to create a plan and updates state. */
-  createPlan(
-    basePlan: BasePlan
-  ): Observable<{ success: boolean; result?: Plan }> {
+  createPlan(basePlan: BasePlan): Observable<{
+    success: boolean;
+    result: Plan;
+  }> {
     return this.createPlanApi(basePlan).pipe(
       take(1),
       map((createdPlan) => {
@@ -125,9 +41,6 @@ export class PlanService {
           success: true,
           result: createdPlan,
         };
-      }),
-      tap(({ result: createdPlan }) => {
-        this.addPlanToState(createdPlan);
       })
     );
   }
@@ -162,8 +75,7 @@ export class PlanService {
       )
       .pipe(
         take(1),
-        map((dbPlan) => this.convertToPlan(dbPlan)),
-        tap((plan) => this.addPlanToState(plan))
+        map((dbPlan) => this.convertToPlan(dbPlan))
       );
   }
 
@@ -198,223 +110,6 @@ export class PlanService {
         withCredentials: true,
       })
       .pipe(take(1));
-  }
-
-  // TODO REMOVE
-  /**
-   * Creates a project in a plan, and returns an ID which can be used to get or update the
-   *  project. "Project" is synonymous with "Config" in the frontend.
-   * */
-  createProjectInPlan(planId: string): Observable<number> {
-    const url = BackendConstants.END_POINT.concat('/planning/create_scenario/');
-    return this.http
-      .post<number>(
-        url,
-        {
-          planning_area: Number(planId),
-        },
-        {
-          withCredentials: true,
-        }
-      )
-      .pipe(take(1));
-  }
-
-  // TODO REMOVE
-  /** Creates project area and returns the ID of the created project area. */
-  createProjectArea(projectId: number, projectArea: GeoJSON.GeoJSON) {
-    const url = BackendConstants.END_POINT.concat('/plan/create_project_area/');
-    return this.http
-      .post<number>(
-        url,
-        {
-          project_id: Number(projectId),
-          geometry: projectArea,
-        },
-        {
-          withCredentials: true,
-        }
-      )
-      .pipe(
-        take(1),
-        catchError((error) => {
-          // TODO: Show error message! Move all error logic to the service.
-          return of(null);
-        })
-      );
-  }
-
-  // TODO REMOVE
-  /** Updates a project with new parameters. */
-  updateProject(projectConfig: ProjectConfig): Observable<number> {
-    const url = BackendConstants.END_POINT.concat('/plan/update_project/');
-    return this.http
-      .put<number>(url, projectConfig, {
-        withCredentials: true,
-      })
-      .pipe(take(1));
-  }
-
-  // TODO REMOVE
-  /** Fetches the projects for a plan from the backend. */
-  getProjectsForPlan(planId: string): Observable<ProjectConfig[]> {
-    const url = BackendConstants.END_POINT.concat(
-      '/planning/list_scenarios_for_planning_area/?planning_area=',
-      planId
-    );
-    return this.http
-      .get(url, {
-        withCredentials: true,
-      })
-      .pipe(
-        take(1),
-        map((response) =>
-          (response as any[]).map((config) =>
-            this.convertToProjectConfig(config)
-          )
-        )
-      );
-  }
-
-  // TODO REMOVE
-  /** Fetches a project by its ID from the backend. */
-  getProject(projectId: number): Observable<ProjectConfig> {
-    const url = BackendConstants.END_POINT.concat(
-      '/planning/get_scenario_by_id/?id=',
-      projectId.toString()
-    );
-    return this.http
-      .get(url, {
-        withCredentials: true,
-      })
-      .pipe(
-        take(1),
-        map((response) => this.convertToProjectConfig(response))
-      );
-  }
-
-  // TODO REMOVE
-  /** Deletes one or more projects from the backend. */
-  deleteProjects(projectIds: number[]): Observable<number[]> {
-    return this.http.post<number[]>(
-      BackendConstants.END_POINT.concat('/plan/delete_projects/'),
-      {
-        project_ids: projectIds,
-      },
-      {
-        withCredentials: true,
-      }
-    );
-  }
-
-  // TODO Add boolean parameter to control if show_results flag is true or false
-  /** Fetches a scenario by its id from the backend. */
-  getScenario(scenarioId: string): Observable<Scenario> {
-    const url = BackendConstants.END_POINT.concat(
-      '/planning/get_scenario_by_id/?id=',
-      scenarioId
-    );
-    return this.http
-      .get(url, {
-        withCredentials: true,
-      })
-      .pipe(
-        take(1),
-        map((response) => this.convertBackendScenarioToScenario(response))
-      );
-  }
-
-  /** Fetches the scenarios for a plan from the backend. */
-  getScenariosForPlan(planId: string): Observable<Scenario[]> {
-    return this.http
-      .get<any[]>(
-        BackendConstants.END_POINT.concat(
-          '/planning/list_scenarios_for_planning_area/?planning_area=',
-          planId
-        ),
-        {
-          withCredentials: true,
-        }
-      )
-      .pipe(
-        map((scenarios) =>
-          scenarios.map(this.convertBackendScenarioToScenario.bind(this))
-        )
-      );
-  }
-
-  /** Creates a scenario in the backend. Returns scenario ID. */
-  createScenario(scenarioParameters: any): Observable<any> {
-    scenarioParameters['configuration'] = this.convertConfigToScenario(
-      scenarioParameters['configuration']
-    );
-    return this.http
-      .post(
-        BackendConstants.END_POINT + '/planning/create_scenario/',
-        scenarioParameters,
-        {
-          withCredentials: true,
-        }
-      )
-      .pipe(take(1));
-  }
-
-  /** Updates a scenario with new notes. */
-  updateScenarioNotes(scenario: Scenario): Observable<number> {
-    const url = BackendConstants.END_POINT.concat('/planning/update_scenario/');
-    return this.http
-      .patch<number>(
-        url,
-        {
-          id: scenario.id,
-          notes: scenario.notes,
-        },
-        {
-          withCredentials: true,
-        }
-      )
-      .pipe(take(1));
-  }
-
-  /** Deletes one or more scenarios from the backend. Returns IDs of deleted scenarios. */
-  deleteScenarios(scenarioIds: string[]): Observable<string[]> {
-    return this.http.post<string[]>(
-      BackendConstants.END_POINT.concat('/planning/delete_scenario/'),
-      {
-        scenario_id: scenarioIds,
-      },
-      {
-        withCredentials: true,
-      }
-    );
-  }
-
-  // TODO REMOVE
-  /** Favorite a scenario in the backend. */
-  favoriteScenario(scenarioId: string): Observable<{ favorited: boolean }> {
-    return this.http.post<{ favorited: boolean }>(
-      BackendConstants.END_POINT.concat('/plan/favorite_scenario/'),
-      {
-        scenario_id: Number(scenarioId),
-      },
-      {
-        withCredentials: true,
-      }
-    );
-  }
-
-  // TODO REMOVE
-  /** Unfavorite a scenario in the backend. */
-  unfavoriteScenario(scenarioId: string): Observable<{ favorited: boolean }> {
-    return this.http.post<{ favorited: boolean }>(
-      BackendConstants.END_POINT.concat('/plan/unfavorite_scenario/'),
-      {
-        scenario_id: Number(scenarioId),
-      },
-      {
-        withCredentials: true,
-      }
-    );
   }
 
   private convertToPlan(plan: BackendPlan): Plan {
@@ -457,206 +152,6 @@ export class PlanService {
       geometry: plan.geometry,
       ownerId: plan.user,
     };
-  }
-
-  private convertToProjectConfig(config: any): ProjectConfig {
-    return {
-      id: config.id,
-      name: config.name,
-      est_cost: config.configuration.est_cost,
-      max_budget: config.configuration.max_budget,
-      min_distance_from_road: config.configuration.min_distance_from_road,
-      max_slope: config.configuration.max_slope,
-      max_treatment_area_ratio: config.configuration.max_treatment_area_ratio,
-      priorities: config.configuration.priorities,
-      weights: config.configuration.weights,
-      excluded_areas: config.configuration.excluded_areas,
-      createdTimestamp: this.convertBackendTimestamptoFrontendTimestamp(
-        config.creation_timestamp
-      ),
-    };
-  }
-
-  private convertToScenarioConfig(config: any): ScenarioConfig {
-    var selectedQuestion: TreatmentQuestionConfig | null = null;
-    this.treatmentGoalsConfig$.subscribe((goals) => {
-      goals!.forEach((goal) => {
-        goal.questions.forEach((question) => {
-          if (
-            question['scenario_priorities']?.toString() ==
-              config.scenario_priorities?.toString() &&
-            question['weights']?.toString() == config.weights?.toString()
-          ) {
-            selectedQuestion = question;
-          }
-        });
-      });
-    });
-
-    return {
-      est_cost: config.est_cost,
-      max_budget: config.max_budget,
-      min_distance_from_road: config.min_distance_from_road,
-      max_slope: config.max_slope,
-      max_treatment_area_ratio: config.max_treatment_area_ratio,
-      treatment_question: selectedQuestion,
-      excluded_areas: config.excluded_areas,
-      stand_size: config.stand_size,
-      created_timestamp: this.convertBackendTimestamptoFrontendTimestamp(
-        config.creation_timestamp
-      ),
-      project_areas: this.convertToProjectAreas(config.project_areas),
-    };
-  }
-
-  private convertBackendScenarioToScenario(scenario: any): Scenario {
-    return {
-      id: scenario.id,
-      name: scenario.name,
-      planning_area: scenario.planning_area,
-      configuration: this.convertToScenarioConfig(scenario.configuration),
-      notes: scenario.notes,
-      scenario_result: scenario.scenario_result,
-    };
-  }
-
-  private convertToProjectAreas(scenarioProjectAreas: {
-    [id: number]: BackendProjectArea;
-  }): ProjectArea[] {
-    if (!scenarioProjectAreas) {
-      return [];
-    }
-
-    let projectAreas: ProjectArea[] = [];
-    Object.values(scenarioProjectAreas).forEach((projectArea) => {
-      projectAreas.push({
-        id: projectArea.id.toString(),
-        projectId: projectArea.properties?.project?.toString(),
-        projectArea: projectArea.geometry,
-        owner: projectArea.properties?.owner?.toString(),
-        estimatedAreaTreated: projectArea.properties?.estimated_area_treated,
-      });
-    });
-
-    return projectAreas;
-  }
-
-  private convertConfigToScenario(config: ScenarioConfig): any {
-    return {
-      est_cost: config.est_cost,
-      max_budget: config.max_budget,
-      min_distance_from_road: config.min_distance_from_road,
-      max_slope: config.max_slope,
-      max_treatment_area_ratio: config.max_treatment_area_ratio,
-      scenario_priorities: config.treatment_question!['scenario_priorities'],
-      scenario_output_fields:
-        config.treatment_question!['scenario_output_fields_paths']!['metrics'],
-      stand_thresholds: config.treatment_question!['stand_thresholds'],
-      global_thresholds: config.treatment_question!['global_thresholds'],
-      weights: config.treatment_question!['weights'],
-      excluded_areas: config.excluded_areas,
-      stand_size: config.stand_size,
-    };
-  }
-
-  // Convert the timestamp stored in backend (measured in seconds since the epoch)
-  // to the timestamp Angular assumes is used (milliseconds since the epoch).
-  convertBackendTimestamptoFrontendTimestamp(
-    timestamp: number | undefined
-  ): number | undefined {
-    return timestamp ? timestamp * 1000 : undefined;
-  }
-
-  private addPlanToState(plan: Plan) {
-    // Object.freeze() enforces shallow runtime immutability
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-        [plan.id]: plan,
-      },
-    });
-
-    this.planState$.next(updatedState);
-  }
-
-  updateStateWithPlan(planId: string | null) {
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-      },
-      currentPlanId: planId,
-    });
-
-    this.planState$.next(updatedState);
-  }
-
-  updateStateWithScenario(scenarioId: string | null) {
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-      },
-      currentScenarioId: scenarioId,
-    });
-
-    this.planState$.next(updatedState);
-  }
-
-  /**
-   * @deprecated
-   */
-  updateStateWithConfig(configId: number | null) {
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-      },
-      currentConfigId: configId,
-    });
-    this.planState$.next(updatedState);
-  }
-
-  updateStateWithConditionLayer(layer: string | null) {
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-      },
-      mapConditionLayer: layer,
-    });
-    this.planState$.next(updatedState);
-  }
-
-  updateStateWithShapes(shapes: any | null) {
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-      },
-      mapShapes: shapes,
-    });
-    this.planState$.next(updatedState);
-  }
-
-  updateStateWithLegendUnits(legendUnits: string | null) {
-    const currentState = Object.freeze(this.planState$.value);
-    const updatedState = Object.freeze({
-      ...currentState,
-      all: {
-        ...currentState.all,
-      },
-      legendUnits: legendUnits,
-    });
-
-    this.planState$.next(updatedState);
   }
 
   private createPlanApi(plan: BasePlan): Observable<Plan> {
@@ -702,46 +197,5 @@ export class PlanService {
         }
       )
       .pipe(take(1));
-  }
-
-  /** Gets Metric Data For Scenario Output Fields */
-  getMetricData(metric_paths: any): Observable<any> {
-    const url = BackendConstants.END_POINT.concat(
-      '/conditions/metrics/?region_name=' +
-        `${regionToString(this.planRegion$.getValue())}` +
-        '&metric_paths=' +
-        JSON.stringify(metric_paths)
-    );
-    return this.http.get<any>(url).pipe(take(1));
-  }
-
-  /**
-   * Updates planRegion and treatmentGoalsConfig if value is a valid Region
-   */
-  setPlanRegion(value: Region) {
-    if (Object.values(Region).includes(value)) {
-      this.planRegion$.next(value);
-      this.http
-        .get<TreatmentGoalConfig[]>(
-          BackendConstants.END_POINT +
-            '/planning/treatment_goals_config/?region_name=' +
-            `${regionToString(this.planRegion$.getValue())}`
-        )
-        .pipe(take(1))
-        .subscribe((config: TreatmentGoalConfig[]) => {
-          this.treatmentGoalsConfig$.next(config);
-        });
-    }
-  }
-
-  downloadCsvData(scenarioId: string): Observable<any> {
-    return this.http.get(
-      BackendConstants.END_POINT +
-        `/planning/get_scenario_download_by_id?id=${scenarioId}`,
-      {
-        withCredentials: true,
-        responseType: 'arraybuffer',
-      }
-    );
   }
 }
