+++ src/interface/src/app/services/treatement-goals.service.spec.ts
@@ -1,0 +1,19 @@
+import { TestBed } from '@angular/core/testing';
+
+import { TreatmentGoalsService } from './treatment-goals.service';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+
+describe('TreatmentGoalsService', () => {
+  let service: TreatmentGoalsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [HttpClientTestingModule],
+    });
+    service = TestBed.inject(TreatmentGoalsService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
