+++ src/interface/src/app/types/index.ts
@@ -5,8 +5,17 @@ export * from './map.types';
 export * from './plan.types';
 export * from './region.types';
 export * from './user.types';
+
 export { regionMapCenters } from '../map/map.helper';
 export { defaultMapConfigsDictionary } from '../map/map.helper';
 export { defaultMapViewOptions } from '../map/map.helper';
 export { defaultMapConfig } from '../map/map.helper';
 export { FrontendConstants } from '../map/map.constants';
+export { ScenarioConfig } from './scenario.types';
+export { Scenario } from './scenario.types';
+export { ScenarioResult } from './scenario.types';
+export { ProjectArea } from './scenario.types';
+export { TreatmentQuestionConfig } from './scenario.types';
+export { ScenarioResultStatus } from './scenario.types';
+export { FeatureCollection } from './scenario.types';
+export { TreatmentGoalConfig } from './scenario.types';
