+++ src/interface/src/app/types/plan.types.ts
@@ -40,89 +40,6 @@ export interface FeatureCollection extends GeoJSON.FeatureCollection {
   properties: any;
 }
 
-export type ScenarioResultStatus =
-  | 'LOADING' // when loading results
-  | 'NOT_STARTED' // Added by FE when the scenario is not created yet.
-  | 'PENDING' // Scenario created, in queue
-  | 'RUNNING' // Scenario created, being processed
-  | 'SUCCESS' // Run completed successfully
-  | 'FAILURE'; // Run failed;
-
-export interface ScenarioResult {
-  status: ScenarioResultStatus;
-  completed_at: string;
-  result: {
-    features: FeatureCollection[];
-    type: string;
-  };
-}
-
-export interface Scenario {
-  id?: string;
-  name: string;
-  notes?: string;
-  planning_area: string;
-  configuration: ScenarioConfig;
-  scenario_result?: ScenarioResult;
-}
-
-export interface ScenarioConfig {
-  created_timestamp?: number;
-  est_cost?: number;
-  max_budget?: number;
-  max_slope?: number;
-  max_treatment_area_ratio?: number;
-  min_distance_from_road?: number;
-  project_areas?: ProjectArea[];
-  treatment_question?: TreatmentQuestionConfig | null;
-  excluded_areas?: string[];
-  stand_size?: string;
-}
-
-export interface TreatmentGoalConfig {
-  category_name?: string;
-  questions: TreatmentQuestionConfig[];
-}
-
-export interface TreatmentQuestionConfig {
-  global_thresholds?: string[];
-  long_question_text?: string;
-  scenario_output_fields_paths?: { [key: string]: string[] };
-  scenario_priorities?: string[];
-  short_question_text?: string;
-  stand_thresholds?: string[];
-  weights?: number[];
-}
-
-export interface Priority {
-  id: string;
-  name: string;
-  weight: number;
-}
-
-export interface ProjectConfig {
-  id: number;
-  name?: string;
-  est_cost?: number;
-  max_budget?: number;
-  max_treatment_area_ratio?: number;
-  min_distance_from_road?: number;
-  max_slope?: number;
-  priorities?: string[];
-  weights?: number[];
-  createdTimestamp?: number;
-  excluded_areas?: { [key: string]: boolean[] };
-}
-
-export interface ProjectArea {
-  id: string;
-  projectId?: string;
-  projectArea: GeoJSON.GeoJSON;
-  owner?: string;
-  estimatedAreaTreated?: number;
-  actualAcresTreated?: number;
-}
-
 export interface PlanConditionScores {
   conditions: PlanConditionScore[];
 }
@@ -130,4 +47,38 @@ export interface PlanConditionScores {
 export interface PlanConditionScore {
   condition: string;
   mean_score: number;
+}
+
+export interface BackendPlan {
+  id?: number;
+  name: string;
+  user?: number;
+  notes?: string;
+  region_name: Region;
+  geometry?: GeoJSON.GeoJSON;
+  scenario_count?: number;
+  projects?: number;
+  created_at?: string;
+  latest_updated?: string;
+}
+
+export interface BackendPlanPreview {
+  id: number;
+  name: string;
+  notes: string;
+  user: number;
+  region_name: Region;
+  scenario_count: number;
+  latest_updated: string;
+  geometry?: GeoJSON.GeoJSON;
+}
+
+export interface BackendProjectArea {
+  id: number;
+  geometry: GeoJSON.GeoJSON;
+  properties?: {
+    estimated_area_treated?: number;
+    owner?: number;
+    project?: number;
+  };
 }
